Z-04 BUILD
The big change is replacement of the old "Grand" default path with a new one called "Garden," which sequences ALL THE PAGES from ALL THE STREAMS in the roughly chrono order they have on Streamlines.  In other words, there's now a quasi-novelistic default reading sequence for VG.  I have edited the Read Me page accordingly.  Also chased down a bunch of layout issues and an erroneous initial setting for pathCount.