ver1.0.1
Enough fixes and improvements for a notch on the build counter.  Added ability to jump into text randomly.  (This is the big one.)  Cards on the "Arcana" page can now be viewed full size.  "Machine" now includes words with links to random pages.  Emojis serve as connectors between "spliced" sentences.